gall: add %rmem task and pill generators
Added the %rmem task to gall. %rmem restores gall to a previously
exported state.

%rmem should be given to gall by unix on boot, to set the state of the
gall apps before they are allowed to start. To facilitate this, we add
+solid-backup and +brass-backup generators. This builds their respective
pills, but appends a %rmem task to the userspace ova. Thus we always
package the backup with a copy of the desk at the time. This is to
ensure that the type of the saved state matches the type of the state
from ++on-save. We need these types to match because we stitch the type
from ++on-save to the raw data from the backup, as we cannot serialise
+type.

%rmem restores the state and incoming and outgoing subscribers. All of
the outgoing subscriptions receive a %kick. For the incoming
subscriptions we call +on-leave.

An %rmem that is received once apps have booted is ignored.